o Corrected bug [1009233] ResultSet getColumnName, getColumnLabel return wrong    values; (thanks to Mike Hutchinson for the fix).  o Added test for bug [1009233] ResultSet getColumnName, getColumnLabel return    wrong values.  o Corrected bug [1015543] ResultSetMetaData.isNullable always returns    columnNoNulls; (thanks to Mike Hutchinson for the bug report and patch).  o Miscellaneous test clean up.
git-svn-id: http://svn.code.sf.net/p/jtds/code/trunk/jtds@531 97a8069c-bbaa-4fa2-9e49-561d8e0b19ef

+++ src/test/net/sourceforge/jtds/test/SAfeTest.java
@@ -900,7 +900,7 @@ public class SAfeTest extends DatabaseTestCase {
         }
 
         pstmt.setBigDecimal(2, maxval);
-        pstmt.setBigDecimal(3, new BigDecimal((double)(1.0 / 3.0))); // Scale > 38
+        pstmt.setBigDecimal(3, new BigDecimal(1.0 / 3.0)); // Scale > 38
         pstmt.setBigDecimal(4, new BigDecimal("12345.56789"));
         assertTrue(pstmt.executeUpdate() == 1);
         pstmt.close();
